cxx-qt-gen: use extern "C++" block to define Qt lib types
