+++ cxx-qt-gen/src/gen_rs.rs
@@ -533,33 +533,6 @@ pub fn generate_qobject_cxx(
                 #[cxx_name = #class_name_str]
                 type #rust_class_name_cpp;
 
-                include!("cxx-qt-lib/include/qt_types.h");
-                #[namespace = ""]
-                type QColor = cxx_qt_lib::QColor;
-                #[namespace = ""]
-                type QDate = cxx_qt_lib::QDate;
-                #[namespace = ""]
-                type QDateTime = cxx_qt_lib::QDateTime;
-                #[namespace = ""]
-                type QPoint = cxx_qt_lib::QPoint;
-                #[namespace = ""]
-                type QPointF = cxx_qt_lib::QPointF;
-                #[namespace = ""]
-                type QRect = cxx_qt_lib::QRect;
-                #[namespace = ""]
-                type QRectF = cxx_qt_lib::QRectF;
-                #[namespace = ""]
-                type QSize = cxx_qt_lib::QSize;
-                #[namespace = ""]
-                type QSizeF = cxx_qt_lib::QSizeF;
-                #[namespace = ""]
-                type QString = cxx_qt_lib::QString;
-                #[namespace = ""]
-                type QTime = cxx_qt_lib::QTime;
-                #[namespace = ""]
-                type QUrl = cxx_qt_lib::QUrl;
-                #[namespace = ""]
-                type QVariant = cxx_qt_lib::QVariant;
                 #update_requester_type
 
                 #(#cpp_functions)*
