+++ cxx-qt-gen/test_inputs/properties.rs
@@ -1,5 +1,9 @@
 mod my_object {
-    use cxx_qt_lib::QColor;
+    #[namespace = ""]
+    unsafe extern "C++" {
+        include!("cxx-qt-lib/include/qt_types.h");
+        type QColor = cxx_qt_lib::QColor;
+    }
 
     #[derive(Default)]
     pub struct Data {
