+++ cxx-qt-gen/test_inputs/signals.rs
@@ -1,5 +1,10 @@
 mod my_object {
-    use cxx_qt_lib::QVariant;
+    #[namespace = ""]
+    unsafe extern "C++" {
+        include!("cxx-qt-lib/include/qt_types.h");
+        type QPoint = cxx_qt_lib::QPoint;
+        type QVariant = cxx_qt_lib::QVariant;
+    }
 
     #[cxx_qt::signals(MyObject)]
     enum MySignals {
