+++ cxx-qt-gen/test_inputs/types_qt_invokable.rs
@@ -1,4 +1,22 @@
 mod my_object {
+    #[namespace = ""]
+    unsafe extern "C++" {
+        include!("cxx-qt-lib/include/qt_types.h");
+        type QColor = cxx_qt_lib::QColor;
+        type QDate = cxx_qt_lib::QDate;
+        type QDateTime = cxx_qt_lib::QDateTime;
+        type QPoint = cxx_qt_lib::QPoint;
+        type QPointF = cxx_qt_lib::QPointF;
+        type QRect = cxx_qt_lib::QRect;
+        type QRectF = cxx_qt_lib::QRectF;
+        type QSize = cxx_qt_lib::QSize;
+        type QSizeF = cxx_qt_lib::QSizeF;
+        type QString = cxx_qt_lib::QString;
+        type QTime = cxx_qt_lib::QTime;
+        type QUrl = cxx_qt_lib::QUrl;
+        type QVariant = cxx_qt_lib::QVariant;
+    }
+
     #[derive(Default)]
     pub struct Data;
