+++ examples/qml_features/src/serialisation.rs
@@ -27,6 +27,12 @@ impl From<Data> for DataSerde {
 mod serialisation {
     use super::DataSerde;
 
+    #[namespace = ""]
+    unsafe extern "C++" {
+        include!("cxx-qt-lib/include/qt_types.h");
+        type QString = cxx_qt_lib::QString;
+    }
+
     pub struct Data {
         pub number: i32,
         pub string: UniquePtr<QString>,
