+++ tests/basic_cxx_qt/src/sub.rs
@@ -6,6 +6,12 @@
 
 #[cxx_qt::bridge]
 pub mod sub_object {
+    #[namespace = ""]
+    unsafe extern "C++" {
+        include!("cxx-qt-lib/include/qt_types.h");
+        type QString = cxx_qt_lib::QString;
+    }
+
     pub struct Data {
         number: i32,
         string: UniquePtr<QString>,
